Move "example data" warning to dashboard pages
Ensemble is almost ready to start hosting real data, even if the
dashboards for that data are pretty simple. So I'd like to start
treating it like a real product. The data warning won't affect all of
our datasets that Ensemble can host, only example datasets, so it should
be shown there.