Don't use index.html to render error pages
There are some benefits to using index.html to handle server errors. For
example, a 404 from the server will cause our pretty "Not Found" page to
be rendered.

However, using index.html to handle server errors also causes some
errors to be invisible. For example, if a new /about page is added and
static.json is not updated with this new path, the /about page will 404
but still render properly. That sucks.

With this change, we won't render pretty 404 pages, but we won't risk
sending a 404 for pages that render properly, either. I think it's worth
it.