Fix #159: Add Firefox and Safari testing envs
I didn't end up adding Opera. There is a node package that wraps the
Opera driver, but it doesn't appear to be especially well-maintained.
Using it would be a liability.

This commit removes the "Markers are hidden upon hover" test. The
moveToElement method doesn't appear to work correctly in Firefox and
Safari. And besides, we're getting rid of that feature soon in #157.