Fix #156: Don't scroll upon region change
This commit ended up being a little more involved than expected. This
bug was happening because the MetricsGraphics component was re-loaded
every time the region selector changed. Fixing the bug meant only
loading the MetricsGraphics component once, which meant responding to
state changes a little more intelligently. So a lot of state-management
code was changed.

The good news is that I took the opportunity to improve performance a
bit. For example, chart descriptions are no longer re-rendered every
time chart data changes.