Fix #178: Update homepage
This commit updates the homepage copy to clarify that a mobile section
of the report is coming soon. The rest of the changes are stylistic. I
didn't want to change the style of the homepage too dramatically, but
some amount of reshuffling was needed for the separate Desktop and
Mobile sections to make sense.