Support running Nightwatch tests on stage and prod
This includes a change to the way we test the <title> element. For some
reason, the old approach was failing with a bizarre error when it was
being used to test the title of the Hardware dashboard in IE using
BrowserStack.

This line:

    browser.assert.title(`Hardware Across the Web | ${browser.globals.siteTitle`);

would fail, with Nightwatch reporting that the actual title was the
literal word "true". This didn't happen when testing locally and or when
testing other dashboards on BrowserStack. I can't imagine what caused
that, but the Hardware dashboard does have the longest title.

In any case, the tests pass when .getTitle() is used. :shrug: