Upgrade Nightwatch to 1.x.x
This is far from perfect. But we don't have many other options.
According to "npm audit", versions of Nightwatch before 1.x.x rely on
many vulnerable dependencies.

This gets us closer to where we need to be. In particular, our use of
WebDriver rather than Selenium is a step in the right direction. There
are some known issues, but I expect them to be resolved over time.

Known issues:

* Tests fail in the "safari" environment when they seemingly should not.
  As a result, the "safari" environment is not run.
* geckodriver cannot be run concurrently. As a result, any environments
  that use geckodriver are not run.
* Automated BrowserStack tests do not work. There is an issue open about
  this. [1]

[1] https://github.com/browserstack/browserstack-integration-nodejs/issues/11